[Cleanup] Moving PJsonFormatter to PCheckerLogJsonFormatter and moving PLogFormatter to PCheckerLogTextFormatter
