+++ Src/PCompiler/CompilerCore/Backend/CSharp/CSharpCodeGenerator.cs
@@ -386,8 +386,6 @@ namespace Plang.Compiler.Backend.CSharp
             context.WriteLine(output);
             context.WriteLine(output, "[PChecker.SystematicTesting.Test]");
             context.WriteLine(output, "public static void Execute(IStateMachineRuntime runtime) {");
-            context.WriteLine(output, "runtime.RegisterLog(new PLogFormatter());");
-            context.WriteLine(output, "runtime.RegisterLog(new PJsonFormatter());");
             context.WriteLine(output, "PModule.runtime = runtime;");
             context.WriteLine(output, "PHelper.InitializeInterfaces();");
             context.WriteLine(output, "PHelper.InitializeEnums();");
